# service-core/docs/swagger.yaml
definitions:
...
  models.DND:
    description: Represents a phone number that should not receive promotional SMS
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      phone_number:
        description: Phone_Number is the number that should not receive promotional
          SMS
        type: string
      reason:
        description: Reason provides the reason for adding the number to the DND list
        type: string
      status:
        description: Status indicates whether the DND entry is active or inactive
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for My Project.
  title: My Project API
  version: "1.0"
paths:
  /api/dnd:
    get:
      consumes:
      - application/json
      description: Get all DND entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DND'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all DND entries
      tags:
      - DND
    post:
      consumes:
      - application/json
      description: Create a new DND entry with phone number, reason, and status
      parameters:
      - description: DND details
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DND'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new DND entry
      tags:
      - DND
  /api/dnd/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a DND entry by ID
      parameters:
      - description: DND ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete an existing DND entry
      tags:
      - DND
    get:
      consumes:
      - application/json
      description: Get details of a DND entry by ID
      parameters:
      - description: DND ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DND'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get DND details
      tags:
      - DND
    put:
      consumes:
      - application/json
      description: 'Update a DND entry by ID with optional fields: phone number, reason,
        and status'
      parameters:
      - description: DND ID
        in: path
        name: id
        required: true
        type: string
      - description: DND details
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DND'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update an existing DND entry
      tags:
      - DND
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"